+++ src/App.js
@@ -9,32 +9,20 @@ let sysInf = store.get("sys");
 let srd = store.get("srd");
 let campaign = store.get("campaign");
 class App extends Component {
-  state = {
-    data: null,
-  };
 
+  
   renderContent = () => {
     if (isMobile) {
+      console.log("rendering mobile");
       return <Mobile srd={srd} sysInf={sysInf} campaign={campaign} />;
     }
+    console.log("rendering desktop");
     return <Desktop srd={srd} sysInf={sysInf} campaign={campaign} />;
   };
 
   componentDidMount() {
-   /*  this.callBackendAPI()
-      .then((res) => this.setState({ data: res.express }))
-      .catch((err) => console.log(err)); */
+    //this.setState({isLoading: false})
   }
-  // fetching the GET route from the Express server which matches the GET route from server.js
-  callBackendAPI = async () => {
-    /* const response = await fetch("/express_backend");
-    const body = await response.json();
-
-    if (response.status !== 200) {
-      throw Error(body.message);
-    }
-    return body; */
-  };
 
   render() {
     return this.renderContent();